GUI Changes: Increased window size, Improved Main Menu screen, Increased map window & added viewpoint centered at extended sewer (and zoomed to a reasonable scale. Listeners for fields when making different selections. Merged Searching by SCI and Predictive Perf. into one window. Controller changes: Merged SCIController and PredictivePerformanceController. Changed task related details to List due to merge. Calculator changes: casted final cost to int in order to drop dec places.

+++ src/main/java/geoview/calculators/MaintenanceScenariosCalculator.java
@@ -38,7 +38,7 @@ public class MaintenanceScenariosCalculator {
             double featureCost = calculateFeatureCost(maintenanceType, feature);
             if((calculatedCost + featureCost) <= budget){
                 calculatedCost += featureCost;
-                feature.put("BUDGET", (Double)featureCost);
+                feature.put("BUDGET", (int)featureCost);
                 budgetedFeatures.add(feature);
             }
         }
